+++ packages/go-kosu/service/cmd.go
@@ -13,9 +13,10 @@ type CommandConfig struct {
 	Web3                 string
 	LiteFullnode         string
 	LAddr                string
+	RemoteSignerListener string
 	RPC                  bool
 	Lite                 bool
-	RemoteSignerListener string
+	Validator            bool
 }
 
 // RegisterCommand register flags and logic into a cobra command to start a node
@@ -23,11 +24,14 @@ func RegisterCommand(cmd *cobra.Command, homeFlag string) {
 	cfg := &CommandConfig{}
 
 	cmd.Flags().StringVarP(&cfg.LAddr, "laddr", "", "tcp://0.0.0.0:26657", "Serve on a given address")
-	cmd.Flags().StringVarP(&cfg.Web3, "web3", "E", "ws://localhost:8546", "URL of an Ethereum JSONRPC provider")
 	cmd.Flags().BoolVarP(&cfg.RPC, "rpc", "", false, "Start the JSON-RPC API")
 	cmd.Flags().BoolVarP(&cfg.Lite, "lite", "", false, "Start the node as a Lite client")
-	cmd.Flags().StringVarP(&cfg.LiteFullnode, "lite-fullnode", "", "http://localhost:26657",
+	cmd.Flags().StringVarP(&cfg.LiteFullnode, "lite.fullnode", "", "http://localhost:26657",
 		"Fullnode's endpoint (required when running with --lite)")
+	cmd.Flags().BoolVarP(&cfg.Validator, "validator", "", false,
+		"Running as a validator node will start the Witness process. This flag is ignored if --lite is enabled")
+	cmd.Flags().StringVarP(&cfg.Web3, "validator.web3", "E", "ws://localhost:8546",
+		"URL of an Ethereum JSONRPC provider")
 
 	// TODO: re-enable this when TM support Tx signing via remote-signer
 	// We can't enable the option without it, because we need to sign Witness's Txs
@@ -55,6 +59,7 @@ func RegisterCommand(cmd *cobra.Command, homeFlag string) {
 			WEB3:                 cfg.Web3,
 			LAddr:                cfg.LAddr,
 			RemoteSignerListener: cfg.RemoteSignerListener,
+			Validator:            cfg.Validator,
 		}
 
 		if cfg.Lite {