+++ packages/go-kosu/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+-   Add --validator family flags to optionally start as a validator node
 -   Add GetBlocks RPC endpoint
 -   Fix wrong RPC remote address in lite mode